+++ lib/core/data_providers/cards_data_provider.dart
@@ -17,6 +17,7 @@ class CardsDataProvider extends ChangeNotifier {
       'MyUCSDChart',
       'student_id',
       'campus_info',
+      'speed_test',
       'staff_id',
       'staff_info',
       'student_info',
@@ -29,7 +30,6 @@ class CardsDataProvider extends ChangeNotifier {
       'events',
       'news',
       'weather',
-      'network_analysis',
     ];
 
     _studentCards = ['student_info', 'student_id', 'finals', 'schedule'];
@@ -84,6 +84,7 @@ class CardsDataProvider extends ChangeNotifier {
           }
           // check to see if card is not active
           else if (!(_availableCards[card].cardActive ?? false)) {
+
             //print()
             _cardOrder.remove(card);
           }
@@ -93,10 +94,12 @@ class CardsDataProvider extends ChangeNotifier {
         for (String card in tempCardStates.keys) {
           // check to see if card no longer exists
           if (_availableCards[card] == null) {
+
             _cardStates.remove(card);
           }
           // check to see if card is not active
           else if (!(_availableCards[card].cardActive ?? false)) {
+
             _cardStates.remove(card);
           }
         }
